condensed logic to single script, updated tsconfig and manifest, and enhanced teamblind blocker removal coverage
