+++ package.json
@@ -2,8 +2,8 @@
     "name": "clear-glass",
     "version": "1.2.0",
     "type": "module",
-    "description": "## The problem",
-    "main": "removeBlockers.js",
+    "description": "Remove blocking elements from popular job review sites",
+    "main": "src/removeBlockers.js",
     "scripts": {},
     "author": "",
     "license": "ISC",
