+++ tsconfig.json
@@ -1,8 +1,10 @@
 {
     "compilerOptions": {
-        "target": "ES6",
-        "module": "ES6",
+        "target": "ES6", // Or whichever version you prefer
+        "module": "CommonJS",
         "strict": true,
-        "esModuleInterop": false
-    }
+        "esModuleInterop": false,
+        "outDir": "./dist"
+    },
+    "include": ["./src"]
 }
